Move stream functions from rregistry to server root
