+++ x-pack/plugins/fleet/server/services/epm/registry/index.ts
@@ -21,7 +21,7 @@ import {
 import { unpackArchiveToCache } from '../archive';
 import { cacheGet, getArchiveFilelist, setArchiveFilelist } from '../archive';
 import { fetchUrl, getResponse, getResponseStream } from './requests';
-import { streamToBuffer } from './streams';
+import { streamToBuffer } from '../../../streams';
 import { getRegistryUrl } from './registry_url';
 import { appContextService } from '../..';
 import { PackageNotFoundError, PackageCacheError } from '../../../errors';
